Altered DB config, added @Entity/@Table/@Column and unique constrains in Bird model, added selection by name, and query plus forbidding two birds with the same name w/ exception error HTTP 500.
Altered configuration for DB, with a few comments and hibernate dialect for Postgres, also the added format_sql. Added @Entity for the creation of table in DB, also added for good practice @Table/@Column with the respective (temporary) names and unique constrain for name so it can't be repeated in the database inside Bird model. Added selection by name in the BirdDao interface, with the SQL query in the BirdDataAccessService for when someone tries to add a bird with the same name to throw an exception HTTP 500 with Message "Bird already registered".