Rewrote the Spring program back code from almost the scratch.
DITCHED jdbc work previously done (since it was difficult to implement)
REWROTE  everything with Spring Data JPA

Functional and working methods working:

- (GET) findAllBirds
- (GET) findById
- (POST) create
- (PUT) update
- (DELETE) delete

Bird data now includes(used wikipedia as reference):
Integer id (UNIQUE)
String commonName
String birdOrder (changed from 'order' because of SQL word)
String family
String genus
String species (UNIQUE)
String photo
String localization (changed from 'location' because of SQL word)

Added custom exceptions:
ObjectNotFoundException - for when the user tries to find or delete a bird by it's Id, but it's and invalid request, returns HTTP 404 and custom message

DuplicatedKeyViolationException - for when the user tries to add a bird with an existing species or when it tries to update a bird with the same species as a different id bird already registered, returns HTTP 400 and custome message.

(Both returns from the ResourceExceptionHandler inside resources.exceptions)

Added a Model Mapper Bean in configuration and changed PostgresDatasource to that package.

BirdService as of now is an interface and the implementation of the methods are extracted to BirdServiceimpl.

Using Bird class now implements Serializable and the (Rest)controller uses a DTO class for communication. From the web (https://www.baeldung.com/java-dto-pattern):

"DTOs or Data Transfer Objects are objects that carry data between processes in order to reduce the number of methods calls. The pattern was first introduced by Martin Fowler in his book EAA.

Fowler explained that the pattern's main purpose is to reduce roundtrips to the server by batching up multiple parameters in a single call. This reduces the network overhead in such remote operations.

Another benefit is the encapsulation of the serialization's logic (the mechanism that translates the object structure and data to a specific format that can be stored and transferred). It provides a single point of change in the serialization nuances. It also decouples the domain models from the presentation layer, allowing both to change independently."

Changed the application.properties for the PostgreSQL and managed to populate the DB by first changing "spring.jpa.hibernate.ddl-auto" to "create" with "sql.init.mode" and "data-locations" commented in at the first run and changing "dll-auto" to "update" and uncommenting the code for "mode" and "data-locations" from the second run.

Was not able to make the initDB.sql work, still in pregress.

Also changed the configuration of the application.properties inside the java test package with working config for H2 memory DB as a future implementation for testing.

As of now, only added test for the main class.