Changed Bird/BirdDto values, Created BirdControllerTest, updated test data in BirdServiceImpl, Updated populateDB.sql with new data vlaues
Changed declared values for Bird and BirdDto so it fits the client necessities and is more in par with the book used in the users Bird Watching as it follows:

    nome português (namePtBr)
    nome inglês (nameEng)
    nome latim (species)
    família (family)
    tamanho (size)
    gênero (gender)
    cor predominante (color)
    habitat (habitat)
    localização (localization)

Created the BirdControllerTest setup for latter implement tests.

Updated the startBird() inside the BirdServiceImpl with the new data for bird.

 Updated the populateDB.sql with the new data format and CHANGED the ID given so it won't conflict with the max number of bird in the book.