Swagger 2 Added for Controller Documentation
Added Swagger 2 (3.0.0) for Documentation for the Controller.

Had to copy a @Bean workaround inside SwaggerConfiguration to 'bypass' Spring Boot default "PathPatter" in conjunction with an 'matching-strategy' as 'ant_path_matcher' inside application properties.

Also added back the configuration for local PostgreSQL (commented) with the configuration for deploy in the cloud for Heroku.

Inside BirdController added lombok for log/doc @Slf4j, the 'log.info' inside the classes and the "ApiOperation" and "ApiReponses" for Swagger documentation in Portuguese from Brazil.

In build.gradle added Lombok also as a annotationProcessor and the needed 'springfox-boot-starter' to wrap it all up with 'swagger2' and 'swagger-ui'.